stash before merge changes
